+++ gltf/src/net/mgsx/gltf/loaders/shared/GLTFLoaderBase.java
@@ -3,6 +3,7 @@ package net.mgsx.gltf.loaders.shared;
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.Camera;
 import com.badlogic.gdx.graphics.Mesh;
+import com.badlogic.gdx.graphics.Pixmap;
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g3d.Material;
 import com.badlogic.gdx.graphics.g3d.Model;
@@ -142,11 +143,9 @@ public class GLTFLoaderBase implements Disposable {
 				imageResolver.load(glModel.images);
 				textureResolver = new TextureResolver();
 				textureResolver.loadTextures(glModel.textures, glModel.samplers, imageResolver);
-				imageResolver.dispose();
 			}
 			
-			materialLoader = new PBRMaterialLoader(textureResolver);
-			// materialLoader = new DefaultMaterialLoader(textureResolver);
+			materialLoader = createMaterialLoader(textureResolver);
 			materialLoader.loadMaterials(glModel.materials);
 			
 			loadCameras();
@@ -163,6 +162,10 @@ public class GLTFLoaderBase implements Disposable {
 			model.scene = scenes.get(glModel.scene);
 			model.maxBones = skinLoader.getMaxBones();
 			model.textures = textureResolver.getTextures(new Array<Texture>());
+			if(imageResolver != null){
+				model.pixmaps = imageResolver.getPixmaps(new Array<Pixmap>());
+				imageResolver.clear();
+			}
 			model.animations = animationLoader.animations;
 			// XXX don't know where the animation are ...
 			for(SceneModel scene : model.scenes){
@@ -177,6 +180,10 @@ public class GLTFLoaderBase implements Disposable {
 			dispose();
 			throw e;
 		}
+	}
+	
+	protected MaterialLoader createMaterialLoader(TextureResolver textureResolver) {
+		return new PBRMaterialLoader(textureResolver);
 	}
 	
 	private void loadLights() {